Move FIFO test into a unix-only test file
The oci publish special-file test used syscall.Mkfifo inline in publish_test.go, which is Unix-specific and could break Windows test compilation. Move it into a build-tagged unix test file and keep the rest of the tests platform-neutral.
