+++ cmd/oci/publish_test.go
@@ -9,7 +9,6 @@ import (
 	"io"
 	"os"
 	"strings"
-	"syscall"
 	"testing"
 )
 
@@ -225,23 +224,6 @@ func TestTarDirSymlinkedEmptyDirPreserved(t *testing.T) {
 	}
 	if !found {
 		t.Fatal("want symlinked empty directory preserved in the tar")
-	}
-}
-
-func TestTarDirSkipsSpecialFile(t *testing.T) {
-	dir := t.TempDir()
-	if err := writeFile(dir+"/ok.txt", "x"); err != nil {
-		t.Fatal(err)
-	}
-	if err := syscall.Mkfifo(dir+"/pipe", 0600); err != nil {
-		t.Skipf("mkfifo unavailable: %s", err)
-	}
-	var buf bytes.Buffer
-	if err := tarDir(dir, &buf); err != nil {
-		t.Fatalf("tarDir should not error or hang on a FIFO: %s", err)
-	}
-	if bytesContains(buf.Bytes(), "pipe") {
-		t.Fatal("want FIFO to be skipped, not written to the tar")
 	}
 }
